Remove completely nonfunctional and misleading plugin
This plugin was added as part of a bulk operation, evidently without the least vetting.

The plugin author never made an effort to make the plugin do what it claims to do; adding arbitrary additional external
source files to the website. They instead simply implementing the copying of the specific file (.htaccess) they needed
for their own application. The plugin author is aware of this deficiency and tracking it in a vague issue report in the
repository, but that has been open for over 2.5 years without any effort to actually implement the promised
functionality.

The description in the catalog and in the plugin documentation claim the plugin provides the arbitrary file copying
capability, and don't make any mention that it doesn't actually do that. Worse, the demonstration snippet in the
documentation appears to configure the plugin to copy a .htaccess file, which would lead the reader who followed the
demo as written when evaluating the plugin to think it provided the promised arbitrary file copying functionality until
such time as they tried to use the plugin in their own application. The only way they could discover why it stopped
working would be to study the plugin's source code or find and correctly interpret the vague issue report in the GitHub
repository.

So clearly the plugin is objectively not qualified for inclusion in the catalog and must be removed out of respect for
the catalog users.